Many updates for uploader, got map and date picker working.
