+++ app/assets/stylesheets/bootstrap-inat.css
@@ -338,3 +338,4 @@ input[type="color"]:focus,
 .modal-container .modal, .modal-container .modal-backdrop {
   position: absolute;
 }
+
