+++ app/webpack/observations/show/components/media_viewer.jsx
@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import Lightbox from "react-images";
 import EasyZoom from "EasyZoom/dist/easyzoom";
 /* global SITE */
