+++ app/webpack/observations/uploader/components/drag_drop_zone.js
@@ -1,5 +1,6 @@
 import _ from "lodash";
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { Grid, Row, Col } from "react-bootstrap";
 import Dropzone from "react-dropzone";
 import { DropTarget } from "react-dnd";
