+++ app/webpack/observations/uploader/components/header_user_menu.js
@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 
 class HeaderUserMenu extends Component {
