+++ app/webpack/observations/uploader/components/left_menu.js
@@ -2,7 +2,7 @@ import _ from "lodash";
 import moment from "moment-timezone";
 import React from "react";
 import PropTypes from "prop-types";
-import { Input, Glyphicon, Accordion, Panel, Badge } from "react-bootstrap";
+import { Glyphicon, PanelGroup, Panel, Badge } from "react-bootstrap";
 import TaxonAutocomplete from "./taxon_autocomplete";
 import DateTimeFieldWrapper from "./date_time_field_wrapper";
 import SelectionBasedComponent from "./selection_based_component";
@@ -50,7 +50,7 @@ class LeftMenu extends SelectionBasedComponent {
     const commonLng = this.commonValue( "longitude" );
     const commonNotes = this.commonValue( "locality_notes" );
     const commonGeoprivacy = this.commonValue( "geoprivacy" );
-    let locationText = commonNotes ||
+    const locationText = commonNotes ||
       ( commonLat && commonLng &&
       `${_.round( commonLat, 4 )},${_.round( commonLng, 4 )}` );
     let multipleGeoprivacy = !commonGeoprivacy && (
@@ -109,7 +109,7 @@ class LeftMenu extends SelectionBasedComponent {
           <input
             type="text"
             className="form-control"
-            value={ commonDate }
+            value={ commonDate || "" }
             onChange= { e => {
               if ( this.refs.datetime ) {
                 this.refs.datetime.onChange( undefined, e.target.value );
@@ -128,7 +128,7 @@ class LeftMenu extends SelectionBasedComponent {
           <input
             type="text"
             className="form-control"
-            value={ locationText }
+            value={ locationText || "" }
             placeholder={ ( this.valuesOf( "latitude" ).length > 1 &&
               this.valuesOf( "longitude" ).length > 1 ) ?
               I18n.t( "edit_multiple_locations" ) : I18n.t( "location" ) }
@@ -144,17 +144,20 @@ class LeftMenu extends SelectionBasedComponent {
             onChange={ e => updateSelectedObsCards( { description: e.target.value } ) }
           />
         </div>
-        <Input
-          key={ `multigeoprivacy${commonGeoprivacy}` }
-          type="select"
-          value={ commonGeoprivacy }
-          onChange={ e => updateSelectedObsCards( { geoprivacy: e.target.value } ) }
-        >
-          { multipleGeoprivacy }
-          <option value="open">{ I18n.t( "location_is_public" ) }</option>
-          <option value="obscured">{ I18n.t( "location_is_obscured" ) }</option>
-          <option value="private">{ I18n.t( "location_is_private" ) }</option>
-        </Input>
+        <div className="form-group">
+          <select
+            key={ `multigeoprivacy${commonGeoprivacy}` }
+            type="select"
+            className="form-control"
+            value={ commonGeoprivacy || "" }
+            onChange={ e => updateSelectedObsCards( { geoprivacy: e.target.value } ) }
+          >
+            { multipleGeoprivacy }
+            <option value="open">{ I18n.t( "location_is_public" ) }</option>
+            <option value="obscured">{ I18n.t( "location_is_obscured" ) }</option>
+            <option value="private">{ I18n.t( "location_is_private" ) }</option>
+          </select>
+        </div>
         <div className="form-group">
           <div className="checkbox">
             <label>
@@ -179,12 +182,12 @@ class LeftMenu extends SelectionBasedComponent {
       <Badge className="count">{ contentCount }</Badge>
     ) : undefined;
     let header = (
-      <div className={ contentCount && "contents" }>
+      <Panel.Title toggle className={ contentCount ? "contents" : undefined }>
         <Glyphicon glyph={ glyph } className="icon" />
         { title }
         <Glyphicon glyph="triangle-right" className={ openGlyphClass } />
         { badge }
-      </div>
+      </Panel.Title>
     );
     let className = `panel-${key}`;
     const onEntered = ( key !== "1" ) ? () => {
@@ -197,12 +200,15 @@ class LeftMenu extends SelectionBasedComponent {
       <Panel
         eventKey={ key }
         className={ className }
-        header={ header }
-        onEnter={ () => { $( `.${className} .toggle` ).addClass( "rotate" ); } }
-        onEntered={ () => setTimeout( onEntered, 50 ) }
-        onExit={ () => { $( `.${className} .toggle` ).removeClass( "rotate" ); } }
       >
-        { contents }
+        <Panel.Heading>
+          { header }
+        </Panel.Heading>
+        <Panel.Collapse onEntered={ () => setTimeout( onEntered, 50 ) }>
+          <Panel.Body>
+            { contents }
+          </Panel.Body>
+        </Panel.Collapse>
       </Panel>
     );
   }
@@ -231,7 +237,7 @@ class LeftMenu extends SelectionBasedComponent {
           />
           <br />
           <br />
-          <Accordion defaultActiveKey="1">
+          <PanelGroup accordion defaultActiveKey="1" id="left-menu-accordion">
             { this.formPanel( "1", I18n.t( "details" ), "pencil",
               this.details( ), detailsContent, true ) }
             { this.formPanel( "2", I18n.t( "tags" ), "tag", (
@@ -240,7 +246,7 @@ class LeftMenu extends SelectionBasedComponent {
               <ProjectsChooser { ...this.props } /> ), projectsContent ) }
             { this.formPanel( "4", I18n.t( "fields_" ), "th-list", (
               <ObservationFieldsChooser { ...this.props } /> ), fieldsContent ) }
-          </Accordion>
+          </PanelGroup>
         </div>
       );
     }
