+++ app/webpack/observations/uploader/components/location_chooser.js
@@ -2,132 +2,24 @@ import _ from "lodash";
 import React from "react";
 import PropTypes from "prop-types";
 import { Modal, Button, Glyphicon } from "react-bootstrap";
-import { GoogleMapLoader, GoogleMap, Circle, SearchBox, Marker,
-  OverlayView } from "react-google-maps";
 import SelectionBasedComponent from "./selection_based_component";
-import util from "../models/util";
-var lastCenterChange = new Date().getTime();
-
-const markerSVG = {
-  path: "M648 1169q117 0 216 -60t156.5 -161t57.5 -218q0 -115 -70 -258q-69 -109 -158 -225.5t-143 " +
-    "-179.5l-54 -62q-9 8 -25.5 24.5t-63.5 67.5t-91 103t-98.5 128t-95.5 148q-60 132 -60 249q0 88 " +
-    "34 169.5t91.5 142t137 96.5t166.5 36zM652.5 974q-91.5 0 -156.5 -65 t-65 -157t65 -156.5t156.5 " +
-    "-64.5t156.5 64.5t65 156.5t-65 157t-156.5 65z",
-  fillOpacity: 1,
-  strokeWeight: 0,
-  scale: 0.02,
-  rotation: 180,
-  origin: new google.maps.Point( 0, 0 ),
-  anchor: new google.maps.Point( 625, 0 )
-};
+import LocationChooserMap from "./location_chooser_map";
 
 class LocationChooser extends SelectionBasedComponent {
 
-  static searchboxStyle( ) {
-    return {
-      border: "1px solid transparent",
-      borderRadius: "2px",
-      boxShadow: "0 2px 6px rgba(0, 0, 0, 0.3)",
-      boxSizing: "border-box",
-      MozBoxSizing: "border-box",
-      fontSize: "13px",
-      height: "31px",
-      marginTop: "10px",
-      outline: "none",
-      padding: "0 12px",
-      textOverflow: "ellipses",
-      width: "250px"
-    };
-  }
-
   constructor( props, context ) {
     super( props, context );
-    this.handleMapClick = this.handleMapClick.bind( this );
     this.close = this.close.bind( this );
     this.save = this.save.bind( this );
-    this.handlePlacesChanged = this.handlePlacesChanged.bind( this );
     this.update = this.update.bind( this );
-    this.fitCircles = this.fitCircles.bind( this );
-    this.reverseGeocode = this.reverseGeocode.bind( this );
-    this.radiusChanged = this.radiusChanged.bind( this );
-    this.centerChanged = this.centerChanged.bind( this );
-    this.moveCircle = this.moveCircle.bind( this );
     this.multiValued = this.multiValued.bind( this );
     this.placeholder = this.placeholder.bind( this );
   }
 
   componentDidUpdate( prevProps ) {
     if ( this.props.show && !prevProps.show ) {
-      if ( !this.props.center ) {
-        setTimeout( this.fitCircles, 10 );
-      }
       // focus on the autocomplete search field when the location modal opens
       setTimeout( () => $( ".gm-style input" ).focus( ).val( "" ), 200 );
-    }
-  }
-
-  fitCircles( ) {
-    if ( !this.refs.map ) { return; }
-    const circles = [];
-    _.each( this.props.obsCards, c => {
-      if ( c.latitude ) {
-        /* global google */
-        circles.push( new google.maps.Circle( {
-          center: {
-            lat: c.latitude,
-            lng: c.longitude
-          },
-          radius: c.accuracy || 0
-        } ) );
-      }
-    } );
-    if ( circles.length > 0 ) {
-      const bounds = new google.maps.LatLngBounds( );
-      _.each( circles, c => {
-        bounds.union( c.getBounds( ) );
-      } );
-      this.refs.map.fitBounds( bounds );
-    }
-  }
-
-  handleMapClick( event ) {
-    const latLng = event.latLng;
-    const zoom = this.refs.map.getZoom( );
-    const radius = Math.round( ( 1 / Math.pow( 2, zoom ) ) * 2000000 );
-    this.moveCircle( latLng, radius, { geocode: true } );
-  }
-
-  moveCircle( center, radius, options = { } ) {
-    this.props.updateState( { locationChooser: {
-      lat: center.lat( ),
-      lng: center.lng( ),
-      center: this.refs.map.getCenter( ),
-      bounds: this.refs.map.getBounds( ),
-      radius
-    } } );
-    if ( options.geocode ) {
-      this.reverseGeocode( center.lat( ), center.lng( ) );
-    }
-  }
-
-  radiusChanged( ) {
-    if ( this.refs.circle ) {
-      const circleState = this.refs.circle.state.circle;
-      this.moveCircle( circleState.center, circleState.radius );
-    }
-  }
-
-  centerChanged( ) {
-    const time = new Date().getTime();
-    if ( time - lastCenterChange > 900 ) {
-      const goTime = time;
-      lastCenterChange = goTime;
-      setTimeout( () => {
-        if ( goTime === lastCenterChange ) {
-          const circleState = this.refs.circle.state.circle;
-          this.moveCircle( circleState.center, circleState.radius, { geocode: true } );
-        }
-      }, 1000 );
     }
   }
 
@@ -140,17 +32,19 @@ class LocationChooser extends SelectionBasedComponent {
       latitude: this.props.lat ? Number( this.props.lat ) : undefined,
       longitude: this.props.lat ? Number( this.props.lng ) : undefined,
       accuracy: this.props.radius ? Number( this.props.radius ) : undefined,
-      center: this.refs.map.getCenter( ),
-      bounds: this.refs.map.getBounds( ),
-      zoom: this.refs.map.getZoom( ),
+      center: this.props.center,
+      bounds: this.props.bounds,
+      zoom: this.props.zoom,
       locality_notes: this.props.notes,
       manualPlaceGuess: this.props.manualPlaceGuess
     };
     if ( !attrs.accuracy ) { attrs.accuracy = undefined; }
     if ( this.props.obsCard ) {
-      this.props.updateSingleObsCard ?
-        this.props.updateObsCard( attrs ) :
+      if ( this.props.updateSingleObsCard ) {
+        this.props.updateObsCard( attrs );
+      } else {
         this.props.updateObsCard( this.props.obsCard, attrs );
+      }
     } else {
       if ( !attrs.latitude && this.multiValued( "latitude" ) ) { delete attrs.latitude; }
       if ( !attrs.longitude && this.multiValued( "longitude" ) ) { delete attrs.longitude; }
@@ -161,76 +55,6 @@ class LocationChooser extends SelectionBasedComponent {
       this.props.updateSelectedObsCards( attrs );
     }
     this.close( );
-  }
-
-  reverseGeocode( lat, lng ) {
-    if ( this.props.manualPlaceGuess && this.props.notes ) { return; }
-    util.reverseGeocode( lat, lng ).then( location => {
-      if ( location ) {
-        this.props.updateState( { locationChooser: {
-          notes: location,
-          manualPlaceGuess: false
-        } } );
-      }
-    } );
-  }
-
-  handlePlacesChanged( ) {
-    const places = this.refs.searchbox.getPlaces();
-    if ( places.length > 0 ) {
-      const geometry = places[0].geometry;
-      const lat = geometry.location.lat( );
-      const lng = geometry.location.lng( );
-      let notes = places[0].formatted_address;
-      let radius;
-      const viewport = geometry.viewport;
-      if ( viewport ) {
-        // radius is the largest distance from geom center to one of the bounds corners
-        radius = _.max( [
-          this.distanceInMeters( lat, lng,
-            viewport.getCenter().lat(), viewport.getCenter().lng() ),
-          this.distanceInMeters( lat, lng,
-            viewport.getNorthEast().lat(), viewport.getNorthEast().lng() )
-        ] );
-        this.refs.map.fitBounds( viewport );
-      } else {
-        notes = this.refs.searchbox.state.inputElement.value || notes;
-        this.refs.map.fitBounds( new google.maps.LatLngBounds(
-          new google.maps.LatLng( lat - 0.001, lng - 0.001 ),
-          new google.maps.LatLng( lat + 0.001, lng + 0.001 ) ) );
-      }
-      let manualPlaceGuess = this.props.manualPlaceGuess;
-      if ( manualPlaceGuess && this.props.notes ) {
-        notes = this.props.notes;
-      } else {
-        manualPlaceGuess = false;
-      }
-      this.props.updateState( { locationChooser: {
-        lat: lat ? lat.toString( ) : undefined,
-        lng: lng ? lng.toString( ) : undefined,
-        center: this.refs.map.getCenter( ),
-        bounds: this.refs.map.getBounds( ),
-        radius,
-        notes,
-        manualPlaceGuess
-      } } );
-    }
-  }
-
-  // Haversine distance calc, adapted from http://www.movable-type.co.uk/scripts/latlong.html
-  distanceInMeters( lat1, lon1, lat2, lon2 ) {
-    const earthRadius = 6370997; // m
-    const degreesPerRadian = 57.2958;
-    const dLat = ( lat2 - lat1 ) / degreesPerRadian;
-    const dLon = ( lon2 - lon1 ) / degreesPerRadian;
-    const lat1Mod = lat1 / degreesPerRadian;
-    const lat2Mod = lat2 / degreesPerRadian;
-
-    const a = Math.sin( dLat / 2 ) * Math.sin( dLat / 2 ) +
-            Math.sin( dLon / 2 ) * Math.sin( dLon / 2 ) * Math.cos( lat1Mod ) * Math.cos( lat2Mod );
-    const c = 2 * Math.atan2( Math.sqrt( a ), Math.sqrt( 1 - a ) );
-    const d = earthRadius * c;
-    return d;
   }
 
   update( field, e ) {
@@ -260,10 +84,6 @@ class LocationChooser extends SelectionBasedComponent {
   }
 
   render() {
-    let center;
-    let circles = [];
-    let markers = [];
-    let overlays = [];
     let canSave = false;
     const latNum = Number( this.props.lat );
     const lngNum = Number( this.props.lng );
@@ -273,66 +93,9 @@ class LocationChooser extends SelectionBasedComponent {
          !_.isNaN( lngNum ) &&
          _.inRange( latNum, -89.999, 90 ) &&
          _.inRange( lngNum, -179.999, 180 ) ) {
-      center = { lat: latNum, lng: lngNum };
       canSave = true;
     } else if ( !this.props.lat && !this.props.lng ) {
       canSave = true;
-    }
-    _.each( this.props.obsCards, c => {
-      if ( c.latitude && !( this.props.obsCard && this.props.obsCard.id === c.id ) ) {
-        const cardImage = $( `[data-id=${c.id}] .carousel-inner img:first` );
-        if ( cardImage.length > 0 ) {
-          overlays.push(
-            <OverlayView key={ `overlay${c.id}` }
-              position={ { lat: c.latitude, lng: c.longitude } }
-              mapPaneName={ OverlayView.OVERLAY_LAYER }
-            >
-              <div className="photo-marker">
-                <img src={ cardImage[0].src } />
-              </div>
-            </OverlayView>
-          );
-        } else {
-          markers.push(
-            <Marker key={`marker${c.id}`}
-              position={{ lat: c.latitude, lng: c.longitude }}
-              icon={ Object.assign( { }, markerSVG, {
-                fillColor: "#333"
-              } ) }
-            />
-          );
-        }
-      }
-    } );
-    if ( center ) {
-      circles.push(
-        <Circle key="circle" ref="circle"
-          center={ center }
-          radius={ Number( this.props.radius ) }
-          onClick={ this.handleMapClick }
-          onRadiusChanged={ this.radiusChanged }
-          onCenterChanged={ this.centerChanged }
-          options={{
-            strokeColor: "#DF0101",
-            strokeOpacity: 0.8,
-            fillColor: "#DF0101",
-            fillOpacity: 0.35
-          }}
-          editable
-          draggable
-        />
-      );
-      if ( !this.props.radius ) {
-        markers.push(
-          <Marker key="marker"
-            position={{ lat: center.lat, lng: center.lng }}
-            icon={ Object.assign( { }, markerSVG, {
-              fillColor: "#DF0101",
-              scale: 0.03
-            } ) }
-          />
-        );
-      }
     }
     const glyph = this.props.notes && ( <Glyphicon glyph="map-marker" /> );
     return (
@@ -349,69 +112,64 @@ class LocationChooser extends SelectionBasedComponent {
           </Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          <GoogleMapLoader
-            containerElement={
-              <div
-                { ...this.props }
-                className="map"
-              />
-            }
-            googleMapElement={
-              <GoogleMap
-                ref="map"
-                defaultZoom={ this.props.zoom || 1 }
-                defaultCenter={ this.props.center || center || { lat: 30, lng: 15 } }
-                onClick={ this.handleMapClick }
-                options={{ streetViewControl: false, fullscreenControl: true }}
-              >
-                <SearchBox
-                  bounds={ this.props.bounds }
-                  onPlacesChanged={ this.handlePlacesChanged }
-                  controlPosition={ google.maps.ControlPosition.TOP_LEFT }
-                  ref="searchbox"
-                  placeholder={ I18n.t( "search_for_a_location" ) }
-                  style={ LocationChooser.searchboxStyle( ) }
-                />
-                { markers }
-                { circles }
-                { overlays }
-              </GoogleMap>
-            }
+          <LocationChooserMap
+            { ...this.props }
+            containerElement={ <div className="map" /> }
+            mapElement={ <div style={{ height: "100%" }} /> }
           />
           <div className="form">
-            <input
-              key="lat"
-              type="text"
-              label={ I18n.t( "latitude" ) }
-              value={ this.props.lat }
-              placeholder={ this.placeholder( "latitude" ) }
-              onChange={ e => this.update( "lat", e ) }
-            />
-            <input
-              key="lng"
-              type="text"
-              label={ I18n.t( "longitude" ) }
-              value={ this.props.lng }
-              placeholder={ this.placeholder( "longitude" ) }
-              onChange={ e => this.update( "lng", e ) }
-            />
-            <input
-              key="radius"
-              type="text"
-              label={ I18n.t( "accuracy_meters" ) }
-              value={ this.props.radius }
-              placeholder={ this.placeholder( "accuracy" ) }
-              onChange={ e => this.update( "radius", e ) }
-            />
-            <input
-              className="notes"
-              key="notes"
-              type="text"
-              label={ I18n.t( "locality_notes" ) }
-              value={ this.props.notes }
-              placeholder={ this.placeholder( "locality_notes" ) }
-              onChange={ e => this.update( "notes", e ) }
-            />
+            <div className="form-group">
+              <label className="control-label">
+                { I18n.t( "latitude" ) }
+                <input
+                  className="form-control"
+                  key="lat"
+                  type="text"
+                  value={ this.props.lat || "" }
+                  placeholder={ this.placeholder( "latitude" ) }
+                  onChange={ e => this.update( "lat", e ) }
+                />
+              </label>
+            </div>
+            <div className="form-group">
+              <label className="control-label">
+                { I18n.t( "longitude" ) }
+                <input
+                  className="form-control"
+                  key="lng"
+                  type="text"
+                  value={ this.props.lng || "" }
+                  placeholder={ this.placeholder( "longitude" ) }
+                  onChange={ e => this.update( "lng", e ) }
+                />
+              </label>
+            </div>
+            <div className="form-group">
+              <label className="control-label">
+                { I18n.t( "accuracy_meters" ) }
+                <input
+                  className="form-control"
+                  key="radius"
+                  type="text"
+                  value={ this.props.radius || "" }
+                  placeholder={ this.placeholder( "accuracy" ) }
+                  onChange={ e => this.update( "radius", e ) }
+                />
+              </label>
+            </div>
+            <div className="form-group">
+              <label className="control-label">
+                { I18n.t( "locality_notes" ) }
+                <input
+                  className="notes form-control"
+                  key="notes"
+                  type="text"
+                  value={ this.props.notes || "" }
+                  placeholder={ this.placeholder( "locality_notes" ) }
+                  onChange={ e => this.update( "notes", e ) }
+                />
+              </label>
+            </div>
           </div>
         </Modal.Body>
         <Modal.Footer>
@@ -431,7 +189,6 @@ class LocationChooser extends SelectionBasedComponent {
 
 LocationChooser.propTypes = {
   show: PropTypes.bool,
-  default: PropTypes.object,
   manualPlaceGuess: PropTypes.bool,
   obsCard: PropTypes.object,
   obsCards: PropTypes.object,
