+++ app/webpack/observations/uploader/components/opening_action_menu.js
@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { Button, DropdownButton, Glyphicon, MenuItem } from "react-bootstrap";
 
 class OpeningActionMenu extends Component {
