+++ app/webpack/observations/uploader/components/photo_viewer.js
@@ -1,5 +1,6 @@
 import _ from "lodash";
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import Lightbox from "react-images";
 
 class PhotoViewer extends Component {
