+++ app/webpack/observations/uploader/components/remove_modal.js
@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import ConfirmModal from "./confirm_modal";
 
 class RemoveModal extends Component {
