+++ app/webpack/observations/uploader/components/sound.jsx
@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from "react";
+import React, { Component } from "react";
+import PropTypes from "prop-types";
 import { DragSource } from "react-dnd";
 import { pipe } from "ramda";
